Move Icon code to seperate module
