+++ lib/index.js
@@ -4,10 +4,9 @@ require('colors')
 
 const errors = require('./errors')
 const fs = require('fs')
-const path = require('path')
 const fontGenerator = require('webfonts-generator')
+const iconGenerator = require('./generateIcons')
 const async = require('async')
-const readline = require('readline')
 
 const CSS_PARSE_REGEX = /\-(.*)\:before.*\n\s*content: "(.*)"/gm
 const FONT_TYPES = [ 'ttf', 'woff', 'woff2', 'eot' ]
@@ -139,74 +138,8 @@ function generate(options, callback) {
 
         fs.writeFile(jsonPath, JSON.stringify(map, null, 4), done)
 
-        //Now generate scss includes
-        let includes = []
-        for(let icon in map) {
-          if(map.hasOwnProperty(icon)) {
-            if (options.classPrefix) {
-              includes.push(`@include makeIcon(${options.classPrefix + "-" + icon}, '${map[icon]}');`)
-            } else {
-              includes.push(`@include makeIcon(${icon}, '${map[icon]}');`)
-            }
-          }
-        }
-
-        //Add css classes for icons
-        const style = `
-[class^="icon-"], [class*=" icon-"] {
-  font-family: 'icons';
-  speak: none;
-  font-style: normal;
-  font-weight: normal;
-  font-variant: normal;
-  text-transform: none;
-
-  /* Better Font Rendering =========== */
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-}`
-
-        fs.writeFile(path.join(outputDir, '/icons.scss'), style + '\n', function (err) {
-          if (err) throw err;
-
-          //Add scss mixin
-          const mixin = `
-@mixin makeIcon($arg, $val) {
-  .#{$arg}:before ,
-  .ion-ios-#{$arg}:before ,
-  .ion-ios-#{$arg}-outline:before ,
-  .ion-md-#{$arg}:before ,
-  .ion-md-#{$arg}-outline:before  {
-    content: $val;
-    font-size: 26px;
-  }
-}
-        `
-          fs.appendFile(path.join(outputDir, '/icons.scss'), mixin + '\n', function (err) {
-            if (err) throw err;
-
-            //Get fonts
-            let lineReader = readline.createInterface({
-              input: require('fs').createReadStream(path.join(outputDir, '/icons.css')),
-            });
-            var lineCounter = 0;
-            var wantedLines = [];
-            lineReader.on('line', function (line) {
-              lineCounter++;
-              wantedLines.push(line);
-              if (lineCounter == 8) { lineReader.close(); }
-            });
-            lineReader.on('close', function () {
-              fs.appendFile(path.join(outputDir, '/icons.scss'), wantedLines.join("\n").replace(new RegExp('./', 'g'), '../assets/fonts/') + '\n', function (err) {
-                if (err) throw err;
-                //Add includes to file
-                for (let include of includes) {
-                  fs.appendFile(path.join(outputDir, '/icons.scss'), include + '\n');
-                }
-              });
-            });
-          })
-        })
+        //start ionicon
+        iconGenerator(map, options)
       } else {
         done()
       }
